remove reference-count related change made redently.
the change made reference count to be increased forever, if we use
listening socket policy.  the change will make it back to the original.

this is not perfect; there will be dangling SA left in the kernel when you
issue "flush" command, for example.  I think this should be tackled very
carefully.