racoon uses syslog() if both the logging file and -F option are not specified when racoon is launched.  Now there are three way to log. One is to the standard output.  Second is to syslog.  Third is to the file specified by the option.
nuke YIPSDEBUG() macro, instead racoon use the value "loglevel"
as logging value.  "loglevel" can be specified either of the following
capital word, and they are printed with each output as the tag name in
the logging file.
	INFO: begin negotiation, SA establishment/deletion/expiration.
	NOTIFY: just notifiable.
	WARNING: not error strictly.
	ERROR: system call error. also invalid parameter/format.
	DEBUG1: debugging informatioin.
	DEBUG2: too more verbose. e.g. parsing config.

XXX The logging level and the tag name are mixed.  They may be separated.
    ERROR should be separated because the error of IKE protocol and system
    call are mixed.
XXX to be more checked.