in defrouter_select, initialized nd_defrouter_primary so that it pointed to the head of the default router list. otherwise, the selection mechanism does not work correctly when the host receives the RA first time, because the link layer address is not stored when the function is called (XXX), and all the entries might be not-probably-reachable.
BTW: I don't think it is a good idea to introduce a separate variable
"nd_defrouter_primary".  In particular, we should not use it to detect
if the default router list is empty.  We can probably find a simple way,
which do not need another state to manage the router selection algorithm.
Can you reconsider the implemenation architecture? > suz