added a new function in6_createmkludge(), which allocated space for multicast kludge at interface initialization time.
Formerly, we dynamically allocated the space in in6_savemkludge() with
malloc(M_WAITOK).  However, it was wrong since the function can be called
under an interrupt context (software timer on address lifetime expiration).
Also, we cannot just give up allocating the strucutre, since the group
membership structure is very complex and we need to keep it anyway.
Of course, this function MUST NOT be called under an interrupt context.
Specifically, it is expected to be called only from in6_ifattach(), though
it is a global function.