Corrected the default router selection algorithm so that round-robin will surely be performed when there's no reachable router.
Also, do not try to install more than one default router, regardless of
the availability of the multipath routing.  This is because I'm not sure about
the benefits for stub hosts comparing to the risk of making the code
complicated and the possibility of introducing bugs.  (the previous code
was actual buggy.)

The bug was found by recent TAHI tests.